+++ Tests/LinuxMain.swift
@@ -1,0 +1,7 @@
+import XCTest
+@testable import EasyInject
+import ValidatedExtensionTests
+
+var tests = [XCTestCaseEntry]()
+tests += ValidatedExtensionTests.ValidatedTests.allTests()
+XCTMain(tests)